added universal mute/unmute
