+++ likeminds-feed-reactnative-integration/components/LMMedia/LMVideo/types.ts
@@ -1,15 +1,16 @@
-import {ReactNode} from 'react';
-import {ViewStyle} from 'react-native';
-import { LMButtonProps } from '../../../uiComponents';
+import { ReactNode } from "react";
+import { ViewStyle } from "react-native";
+import { LMButtonProps } from "../../../uiComponents";
 
 export interface LMVideoProps {
   videoUrl: string; // url of the video to be displayed
   height?: number; // height of the video player
   width?: number; // width of the video player
   videoStyle?: ViewStyle; // this represents the style of the video player
-  boxFit?: 'stretch' | 'contain' | 'cover' | 'none'; // this represents how the video player should be fitted in its wrapper view
+  boxFit?: "stretch" | "contain" | "cover" | "none"; // this represents how the video player should be fitted in its wrapper view
   boxStyle?: ViewStyle; // this represents the style of the view that contains the video player
   aspectRatio?: 0 | 0.1 | 0.2 | 0.3 | 0.4 | 0.5 | 0.6 | 0.7 | 0.8 | 0.9 | 1; // ratio of the video player between 0 to 1
+  showMuteUnmute?: boolean;
   showControls?: boolean; // this represents if the controls of the player has to be shown or not
   looping?: boolean; // this represents if the video should play on loop or not
   loaderWidget?: ReactNode; // this represents the loader component
@@ -23,5 +24,5 @@ export interface LMVideoProps {
   postId?: string;
   videoInFeed?: boolean;
   videoInCarousel?: boolean;
-  currentVideoInCarousel?: string
+  currentVideoInCarousel?: string;
 }
