+++ likeminds-feed-reactnative-integration/screens/carouselScreen/index.tsx
@@ -1,11 +1,4 @@
-import {
-  View,
-  Text,
-  Image,
-  TouchableOpacity,
-  BackHandler,
-  ActivityIndicator,
-} from "react-native";
+import { View, Text, Image, TouchableOpacity, BackHandler } from "react-native";
 import React, { useEffect, useRef, useState } from "react";
 import { SwiperFlatList } from "react-native-swiper-flatlist";
 import Layout from "../../constants/Layout";
@@ -22,34 +15,12 @@ import STYLES from "../../constants/Styles";
 import { STATUS_BAR_STYLE } from "../../store/types/types";
 import { useAppDispatch } from "../../store/store";
 import { ReactNativeZoomableView } from "@openspacelabs/react-native-zoomable-view";
-import Video from "react-native-video";
-import Slider from "@react-native-community/slider";
+import { VideoPlayer } from "../../components";
 
 const CarouselScreen = ({ navigation, route }: any) => {
-  const video = useRef<any>(null);
   const dispatch = useAppDispatch();
   const { index, dataObject, backIconPath } = route.params;
   const data = dataObject?.attachments;
-
-  const ref = useRef<any>();
-  const [clicked, setClicked] = useState(false);
-  const [puased, setPaused] = useState(false);
-  const [progress, setProgress] = useState<any>(null);
-  const [mute, setMute] = useState(false);
-  const format = (seconds) => {
-    /* @ts-ignore */
-    let mins = parseInt(seconds / 60)
-      .toString()
-      .padStart(2, "0");
-    let secs = (Math.trunc(seconds) % 60).toString().padStart(2, "0");
-    return `${mins}:${secs}`;
-  };
-
-  const resetClicked = () => {
-    setTimeout(() => {
-      setClicked(false);
-    }, 3000);
-  };
 
   let imageCount = 0;
   let videoCount = 0;
@@ -214,139 +185,7 @@ const CarouselScreen = ({ navigation, route }: any) => {
                   />
                 </ReactNativeZoomableView>
               ) : item?.attachmentType === VIDEO_ATTACHMENT_TYPE ? (
-                <View style={styles.video}>
-                  <TouchableOpacity
-                    style={styles.video}
-                    onPress={() => {
-                      if (progress !== null) {
-                        setClicked(true);
-                        resetClicked(); // Reset clicked after 3 seconds
-                      }
-                    }}
-                  >
-                    {progress == null && (
-                      <View style={styles.activityIndicatorContainer}>
-                        <ActivityIndicator
-                          size="large"
-                          color={STYLES.$COLORS.SECONDARY}
-                          style={{
-                            justifyContent: "center",
-                            alignItems: "center",
-                            zIndex: 2,
-                          }}
-                        />
-                      </View>
-                    )}
-                    <Video
-                      paused={puased}
-                      source={{
-                        uri: item?.attachmentMeta?.url,
-                      }}
-                      ref={ref}
-                      onProgress={(x) => {
-                        setProgress(x);
-                      }}
-                      style={styles.videoPlayer}
-                      resizeMode="contain"
-                      muted={mute}
-                      onEnd={() => {
-                        setPaused(true); // Pause the video
-                        setProgress({ ...progress, currentTime: 0 }); // Reset seek position
-                        ref.current.seek(0); // Seek to the beginning of the video
-                      }}
-                    />
-                    {clicked && (
-                      <TouchableOpacity
-                        style={{
-                          width: "100%",
-                          height: "100%",
-                          position: "absolute",
-                          backgroundColor: "rgba(0,0,0,.5)",
-                          justifyContent: "center",
-                          alignItems: "center",
-                        }}
-                      >
-                        <View style={{ flexDirection: "row" }}>
-                          <TouchableOpacity
-                            onPress={() => {
-                              setPaused(!puased);
-                            }}
-                          >
-                            <Image
-                              source={
-                                puased
-                                  ? require("../../assets/images/play-button.png")
-                                  : require("../../assets/images/pause-button.png")
-                              }
-                              style={{
-                                width: 40,
-                                height: 40,
-                                tintColor: "white",
-                              }}
-                            />
-                          </TouchableOpacity>
-                        </View>
-                        <View
-                          style={{
-                            width: "100%",
-                            flexDirection: "row",
-                            justifyContent: "space-between",
-                            position: "absolute",
-                            bottom: 0,
-                            paddingLeft: 10,
-                            paddingRight: 50,
-                            alignItems: "center",
-                          }}
-                        >
-                          <Text style={{ color: "white" }}>
-                            {format(progress.currentTime)}
-                          </Text>
-                          <Slider
-                            style={{ width: "80%", height: 40 }}
-                            minimumValue={0}
-                            maximumValue={progress.seekableDuration}
-                            minimumTrackTintColor="#FFFFFF"
-                            maximumTrackTintColor="#FFFFFF"
-                            step={0}
-                            value={progress.currentTime}
-                            onValueChange={(x) => {
-                              ref.current.seek(x);
-                            }}
-                          />
-                          <Text style={{ color: "white", marginRight: 10 }}>
-                            {format(progress.seekableDuration)}
-                          </Text>
-                          <TouchableOpacity onPress={() => setMute(!mute)}>
-                            <Image
-                              source={
-                                mute
-                                  ? require("../../assets/images/muted.png")
-                                  : require("../../assets/images/unmute.png")
-                              }
-                              style={{
-                                height: mute ? 20 : 25,
-                                width: mute ? 20 : 25,
-                                tintColor: "white",
-                              }}
-                            />
-                          </TouchableOpacity>
-                        </View>
-                        <View
-                          style={{
-                            width: "100%",
-                            flexDirection: "row",
-                            justifyContent: "space-between",
-                            position: "absolute",
-                            top: 10,
-                            paddingLeft: 20,
-                            paddingRight: 20,
-                            alignItems: "center",
-                          }}
-                        ></View>
-                      </TouchableOpacity>
-                    )}
-                  </TouchableOpacity>
-                </View>
+                <VideoPlayer url={item?.attachmentMeta?.url} />
               ) : null}
             </View>
           );
