+++ likeminds-feed-reactnative-integration/store/types/types.ts
@@ -7,6 +7,7 @@ export const MEMBER_STATE_SUCCESS = "MEMBER_STATE_SUCCESS";
 export const MEMBER_STATE_FAILED = "MEMBER_STATE_FAILED";
 
 export const STATUS_BAR_STYLE = "STATUS_BAR_STYLE";
+export const SET_MUTED_STATE = "SET_MUTED_STATE";
 
 export const UNIVERSAL_FEED_DATA = "UNIVERSAL_FEED_DATA";
 export const UNIVERSAL_FEED_FAILED = "UNIVERSAL_FEED_FAILED";
